+++ src/AbstractSinglyLinkedList.php
@@ -22,9 +22,6 @@ use Obernard\LinkedList\Exception\NodeException;
  * 
  * The link between two nodes is an unidirectionnal relation:
  * 
- * lpushn() and lpopn() methods makes AbstractSinglyLinkedList similar
- *  to first-in/last-out queue-lists.  
- * 
  * 
  * @author Olivier Bernard
  */